// src/triage/stage.enum.ts
export enum PipelineStage {
  INIT = 'INIT',
  INPUTS_LOADED = 'INPUTS_LOADED',
  TICKETS_NORMALIZED = 'TICKETS_NORMALIZED',
  TRIAGE_PREDICTED = 'TRIAGE_PREDICTED',
  HUMAN_REVIEW_COMPLETE = 'HUMAN_REVIEW_COMPLETE',
  FINAL_QUEUE_GENERATED = 'FINAL_QUEUE_GENERATED',
  VALIDATION_COMPLETE = 'VALIDATION_COMPLETE',
  RESULTS_FINALISED = 'RESULTS_FINALISED',
}

export const STAGE_ORDER: PipelineStage[] = [
  PipelineStage.INIT,
  PipelineStage.INPUTS_LOADED,
  PipelineStage.TICKETS_NORMALIZED,
  PipelineStage.TRIAGE_PREDICTED,
  PipelineStage.HUMAN_REVIEW_COMPLETE,
  PipelineStage.FINAL_QUEUE_GENERATED,
  PipelineStage.VALIDATION_COMPLETE,
  PipelineStage.RESULTS_FINALISED,
];